requirements.txt
configparser==3.7.4
lxml==4.4.0
numpy==1.17.0
ObjectListView==1.3.1
Pillow==6.1.0
PyPubSub==3.3.0
python-dateutil==2.8.0
six==1.12.0
wxPython==4.0.6
